Always log the libclang version when generating bindings
Before, we would only log the libclang version when running bindgen on the
command line. Now we do it every time bindings are generated. This is super
useful information to have when debugging bindgen invocations.
