+++ src/main.rs
@@ -7,13 +7,8 @@ extern crate log;
 extern crate clang_sys;
 extern crate clap;
 
-use bindgen::clang_version;
 use std::env;
 use std::panic;
-
-#[macro_use]
-#[cfg(not(feature="logging"))]
-mod log_stubs;
 
 mod options;
 use options::builder_from_flags;
@@ -29,26 +24,6 @@ pub fn main() {
         .expect("Failed to set logger.");
 
     let bind_args: Vec<_> = env::args().collect();
-
-    let version = clang_version();
-    let expected_version = if cfg!(feature = "testing_only_libclang_4") {
-        (4, 0)
-    } else if cfg!(feature = "testing_only_libclang_3_8") {
-        (3, 8)
-    } else {
-        // Default to 3.9.
-        (3, 9)
-    };
-
-    info!("Clang Version: {}", version.full);
-
-    match version.parsed {
-        None => warn!("Couldn't parse libclang version"),
-        Some(version) if version != expected_version => {
-            warn!("Using clang {:?}, expected {:?}", version, expected_version);
-        }
-        _ => {}
-    }
 
     match builder_from_flags(bind_args.into_iter()) {
         Ok((builder, output, verbose)) => {
